+++ Editor/Utils/EditorUtils_Properties.cs
@@ -50,8 +50,11 @@ namespace Vertx.Utilities.Editor
 			Debug.Log(stringBuilder);
 		}
 
-		public static SerializedProperty FindBackingProperty(this SerializedProperty property, string propertyName) =>
+		public static SerializedProperty FindBackingPropertyRelative(this SerializedProperty property, string propertyName) =>
 			property.FindPropertyRelative($"<{propertyName}>k__BackingField");
+		
+		public static SerializedProperty FindBackingProperty(this SerializedObject property, string propertyName) =>
+			property.FindProperty($"<{propertyName}>k__BackingField");
 
 		private const int safetySerializationDepth = 10;